+++ docs/source/conf.py
@@ -35,10 +35,11 @@ extensions = [
     'sphinx.ext.autosectionlabel',
     'sphinx.ext.viewcode',
     'sphinx.ext.napoleon',
-    'sphinx.ext.mathjax'
+    'sphinx.ext.mathjax',
+    'sphinx_rtd_theme'
 ]
-
-autodoc_class_signature = "separated"
+autoclass_content = 'both'
+#autodoc_class_signature = "separated"
 
 # Make sure the target is unique
 autosectionlabel_prefix_document = True