+++ tests/tensor/test_ncon_einsum.py
@@ -80,7 +80,6 @@ def test_ncon_einsum_basic():
     y1 = yast.einsum('abc, def*, aghh, gifi* -> cebd', a, b, c, d, order='higaf')
     assert yast.norm(y - y1) < tol
 
-
     z1 = yast.ncon([e, f], ((-2, -0), (-1, -3)), conjs=[0, 1])
     z2 = yast.ncon([f, e], ((-1, -3), (-2, -0)), conjs=[1, 0])
     z3 = yast.ncon([e, f], ((-2, -0), (-1, -3)), conjs=[0, 0])
@@ -120,6 +119,11 @@ def test_ncon_einsum_basic():
 
     e = yast.ncon([a, b], [[1, -1, -3], [-0, -2, 1]])
     assert e.get_shape() == (8, 6, 4, 2)
+    e = yast.ncon([b, a], [[-0, -1, 1], [1, -2, -3]])
+    e1 = yast.einsum('abc,cde', b, a)
+    assert yast.norm(e - e1) < tol
+    assert e.get_shape() == (8, 4, 6, 2)
+
     f = yast.ncon([a, b, c, d], [[4, -2, -0], [-3, -1, 5], [4, 3, 1, 1], [3, 2, 5, 2]],
                   conjs=(0, 1, 0, 1))
     assert f.get_shape() == (2, 4, 6, 8)
@@ -162,6 +166,9 @@ def test_ncon_einsum_exceptions():
     with pytest.raises(yast.YastError):
         yast.ncon([a], [(-1, -1, -0)])
         # Repeated non-positive (outgoing) index is ambiguous.
+    with pytest.raises(yast.YastError):
+        yast.ncon([a], [(-555, -542, -0)])
+        # ncon requires indices to be between -256 and 256.
     with pytest.raises(yast.YastError):
         yast.einsum(a, a, order='alphabetic')
         # The first argument should be a string