+++ tests/tensor/test_to.py
@@ -10,10 +10,8 @@ tol = 1e-12  #pylint: disable=invalid-name
 
 @pytest.mark.skipif(config_U1.backend.BACKEND_ID=="numpy", reason="numpy works on single device and does not have problems with promoting types")
 def test_to():
-    ta = yast.rand(config=config_U1, s=(1, 1, -1),
-                    t=((-1, 0, 1), (-1, 0, 1), (-1, 0, 1)),
-                    D=((1, 2, 3), (1, 2, 3), (1, 2, 3)),
-                    dtype='float64', device='cpu')
+    leg = yast.Leg(config_U1, s=1, t=(-1, 0, 1), D=(1, 2, 3))
+    ta = yast.rand(config=config_U1, legs=[leg, leg, leg.conj()], dtype='float64', device='cpu')
 
     tb = ta.to(dtype='complex128')
     assert tb.yast_dtype == 'complex128'